Avoid overflow when calculating unused_memory_limit_in_bytes.
size_t may be 32-bit on 32-bit systems. Need to use int64
in calculation.

BUG=229427


Review URL: https://chromiumcodereview.appspot.com/13841009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193172 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9d52c1c3a7387664d91ca3863f13d26a8a8aafc6

+++ cc/trees/layer_tree_host_impl.cc
@@ -903,9 +903,9 @@ void LayerTreeHostImpl::EnforceManagedMemoryPolicy(
                                       policy.bytes_limit_when_not_visible;
     // TODO(reveman): We should avoid keeping around unused resources if
     // possible. crbug.com/224475
-    new_state.unused_memory_limit_in_bytes =
-        (new_state.memory_limit_in_bytes *
-         settings_.max_unused_resource_memory_percentage) / 100;
+    new_state.unused_memory_limit_in_bytes = static_cast<size_t>(
+        (static_cast<int64>(new_state.memory_limit_in_bytes) *
+         settings_.max_unused_resource_memory_percentage) / 100);
     new_state.memory_limit_policy =
         ManagedMemoryPolicy::PriorityCutoffToTileMemoryLimitPolicy(
             visible_ ?
